Improve DYFI charts code, move DYFI service
