+++ src/app/core/core.module.ts
@@ -9,7 +9,6 @@ import { FormatterService } from './formatter.service';
 import { GeoserveService } from './geoserve.service';
 import { StationService } from './station.service';
 import { MetadataService } from './metadata.service';
-import { DyfiService } from './dyfi.service';
 
 @NgModule({
   imports: [
@@ -29,8 +28,7 @@ export class CoreModule {
         GeoserveService,
         QuakemlService,
         StationService,
-        MetadataService,
-        DyfiService
+        MetadataService
       ]
     };
   }
