+++ src/app/dyfi/dyfi.module.ts
@@ -9,10 +9,11 @@ import {
 
 import { DyfiComponent } from './dyfi/dyfi.component';
 import { DyfiRoutingModule } from './dyfi-routing.module';
+import { DyfiService } from './dyfi.service';
 import { IntensityComponent } from './intensity/intensity.component';
 import { IntensityVsDistanceComponent } from './intensity-vs-distance/intensity-vs-distance.component';
-import { ProductPageModule } from '../product-page/product-page.module';
 import { NgxChartsModule } from '../shared/ngx-charts/ngx-charts.module';
+import { ProductPageModule } from '../product-page/product-page.module';
 import { ResponsesVsTimeComponent } from './responses-vs-time/responses-vs-time.component';
 import { ResponsesComponent } from './responses/responses.component';
 import { ZipComponent } from './zip/zip.component';
@@ -36,6 +37,9 @@ import { ZipComponent } from './zip/zip.component';
     IntensityVsDistanceComponent,
     ResponsesVsTimeComponent,
     ResponsesComponent
+  ],
+  providers: [
+    DyfiService
   ]
 })
 export class DyfiModule { }
