+++ src/app/dyfi/intensity-vs-distance/intensity-vs-distance.component.spec.ts
@@ -3,7 +3,7 @@ import { async, ComponentFixture, TestBed } from '@angular/core/testing';
 import { MockComponent } from 'ng2-mock-component';
 import { of } from 'rxjs/observable/of';
 
-import { DyfiService } from '../../core/dyfi.service';
+import { DyfiService } from '../dyfi.service';
 import { EventService } from '../../core/event.service';
 import { IntensityVsDistanceComponent } from './intensity-vs-distance.component';
 
@@ -106,7 +106,8 @@ describe('IntensityVsDistanceComponent', () => {
     it('handles null data', () => {
       component.onDyfiSeries(null);
 
-      expect(component.dyfiSeries).toBe(null);
+      expect(component.bubbleSeries).toBe(null);
+      expect(component.lineSeries).toBe(null);
     });
 
     it('handles known data', () => {
