+++ src/app/dyfi/responses-vs-time/responses-vs-time.component.spec.ts
@@ -3,7 +3,7 @@ import { async, ComponentFixture, TestBed } from '@angular/core/testing';
 import { MockComponent } from 'ng2-mock-component';
 import { of } from 'rxjs/observable/of';
 
-import { DyfiService } from '../../core/dyfi.service';
+import { DyfiService } from '../dyfi.service';
 import { EventService } from '../../core/event.service';
 import { ResponsesVsTimeComponent } from './responses-vs-time.component';
